Servant.JS.Angular

Documentation

data AngularOptions

Options specific to the angular code generator

Constructors

AngularOptions

Fields

serviceName :: Text

When generating code with wrapInService, name of the service to generate

prologue :: Text -> Text -> Text

beginning of the service definition

epilogue :: Text

end of the service definition

defAngularOptions :: AngularOptions

Default options for the Angular codegen. Used by wrapInService.

angularService :: AngularOptions -> JavaScriptGenerator

Instead of simply generating top level functions, generates a service instance on which your controllers can depend to access your API. This variant uses default AngularOptions.

angularServiceWith :: AngularOptions -> CommonGeneratorOptions -> JavaScriptGenerator

Instead of simply generating top level functions, generates a service instance on which your controllers can depend to access your API

angular :: AngularOptions -> JavaScriptGenerator

Generate regular javacript functions that use the $http service, using default values for CommonGeneratorOptions.

angularWith :: AngularOptions -> CommonGeneratorOptions -> JavaScriptGenerator

Generate regular javascript functions that use the $http service.

generateAngularJS :: AngularOptions -> AjaxReq -> Text

js codegen using $http service from Angular using default options

generateAngularJSWith :: AngularOptions -> CommonGeneratorOptions -> AjaxReq -> Text

js codegen using $http service from Angular