Servant.JS.Internal

Documentation

type JavaScriptGenerator = [Req] -> Text

data CommonGeneratorOptions

This structure is used by specific implementations to let you customize the output

Constructors

CommonGeneratorOptions

Fields

functionNameBuilder :: FunctionName -> Text

function generating function names

requestBody :: Text

name used when a user want to send the request body (to let you redefine it)

successCallback :: Text

name of the callback parameter when the request was successful

errorCallback :: Text

name of the callback parameter when the request reported an error

moduleName :: Text

namespace on which we define the foreign function (empty mean local var)

urlPrefix :: Text

a prefix we should add to the Url in the codegen

defCommonGeneratorOptions :: CommonGeneratorOptions

Default options.

> defCommonGeneratorOptions = CommonGeneratorOptions
>   { functionNameBuilder = camelCase
>   , requestBody = "body"
>   , successCallback = "onSuccess"
>   , errorCallback = "onError"
>   , moduleName = ""
>   , urlPrefix = ""
>   }

type AjaxReq = Req

jsSegments :: [Segment] -> Text

segmentToStr :: Segment -> Bool -> Text

segmentTypeToStr :: SegmentType -> Text

jsParams :: [QueryArg] -> Text

jsGParams :: Text -> [QueryArg] -> Text

paramToStr :: QueryArg -> Bool -> Text

toValidFunctionName :: Text -> Text

Attempts to reduce the function name provided to that allowed by Foreign.

https://mathiasbynens.be/notes/javascript-identifiers Couldn't work out how to handle zero-width characters.

@TODO: specify better default function name, or throw error?

toJSHeader :: HeaderArg -> Text

data a :<|> b :: * -> * -> * infixr 8

Union of two APIs, first takes precedence in case of overlap.

Example:

>>> :{
type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
       :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] () -- POST /books
:}

Constructors

a :<|> b infixr 8

data path :> a :: k -> k1 -> * infixr 9

The contained API (second argument) can be found under ("/" ++ path) (path being the first argument).

Example:

>>> -- GET /hello/world
>>> -- returning a JSON encoded World value
>>> type MyApi = "hello" :> "world" :> Get '[JSON] World

defReq :: Req

reqHeaders :: Lens' Req [HeaderArg]

class HasForeign layout where

Associated Types

type Foreign layout :: *

Methods

foreignFor :: Proxy * layout -> Req -> Foreign layout

data HeaderArg :: *

Constructors

HeaderArg

Fields

headerArgName :: Text

ReplaceHeaderArg

Fields

headerArgName :: Text

headerPattern :: Text

concatCase :: FunctionName -> Text

Function name builder that simply concat each part together

snakeCase :: FunctionName -> Text

Function name builder using the snake_case convention. each part is separated by a single underscore character.

camelCase :: FunctionName -> Text

Function name builder using the CamelCase convention. each part begins with an upper case character.

data ReqBody contentTypes a :: [*] -> k -> *

Extract the request body as a value of type a.

Example:

>>> -- POST /books
>>> type MyApi = "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book

data JSON :: *

data FormUrlEncoded :: *

data Post contentTypes a :: [*] -> * -> *

Endpoint for POST requests. The type variable represents the type of the response body (not the request body, use ReqBody for that).

Example:

>>> -- POST /books
>>> -- with a JSON encoded Book as the request body
>>> -- returning the just-created Book
>>> type MyApi = "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book

data Get contentTypes a :: [*] -> * -> *

Endpoint for simple GET requests. Serves the result as JSON.

Example:

>>> type MyApi = "books" :> Get '[JSON] [Book]

data Raw :: *

Endpoint for plugging in your own Wai Applications.

The given Application will get the request as received by the server, potentially with a modified (stripped) pathInfo if the Application is being routed with :>.

In addition to just letting you plug in your existing WAI Applications, this can also be used with serveDirectory to serve static files stored in a particular directory on your filesystem

data Header sym a :: Symbol -> * -> *

Extract the given header's value as a value of type a.

Example:

>>> newtype Referer = Referer Text deriving (Eq, Show)
>>> 
>>> -- GET /view-my-referer
>>> type MyApi = "view-my-referer" :> Header "from" Referer :> Get '[JSON] Referer